requirements.txt
py-clob-client>=0.12.0
websockets>=12.0
numpy>=1.26.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
